+++ src/main/kotlin/org/ifinalframework/plugins/aio/common/util/UAnnotatedKt.kt
@@ -1,0 +1,17 @@
+package org.ifinalframework.plugins.aio.common.util
+
+import org.jetbrains.uast.UAnnotated
+import org.jetbrains.uast.UAnnotation
+
+
+/**
+ * UAnnotatedKt
+ *
+ * @author iimik
+ * @since 0.0.5
+ **/
+object UAnnotatedKt {
+    fun UAnnotated.findAnyAnnotation(vararg anns: String): UAnnotation? {
+        return anns.firstNotNullOfOrNull { this.findAnnotation(it) }
+    }
+}