+++ src/test/kotlin/org/ifinalframework/plugins/aio/spring/feign/FeignLineMarkerClient.kt
@@ -1,0 +1,18 @@
+package org.ifinalframework.plugins.aio.spring.feign
+
+import org.springframework.cloud.openfeign.FeignClient
+import org.springframework.web.bind.annotation.GetMapping
+
+
+/**
+ * FeignLineMarkerClient
+ *
+ * @author iimik
+ * @since 0.0.5
+ **/
+@FeignClient(path = "/api/feign")
+interface FeignLineMarkerClient {
+
+    @GetMapping("/get")
+    fun getFeign() 
+}